Add files via upload
This assignment takes place from the "ServiceNow" website.  This scenario,  similar  to how an  IT Support Specialist or Systems administrator  would solve a problem  in  their job would  require  me to  first  identify an issue, and then begin the necessary steps in order to adjust this issue. As if I were assisting a customer with a technical  problem .