Identifying List and Filters (Part Two)
In the second portion of this assignment,  I navigated through the menu then  I located the  "Show/hide" filter (funnel) icon.  After that step, I set the field, operator(filter/condition) and selected these options : " State | is not | Closed AND
Assignment Group | is not empty AND
Assigned to | is not empty".